Move 2 lexers to distinct files
