+++ lexers/g/gosu.go
@@ -17,16 +17,3 @@ var Gosu = internal.Register(MustNewLexer(
 		"root": {},
 	},
 ))
-
-// GosuTemplate lexer.
-var GosuTemplate = internal.Register(MustNewLexer(
-	&Config{
-		Name:      "Gosu Template",
-		Aliases:   []string{"gst"},
-		Filenames: []string{"*.gst"},
-		MimeTypes: []string{"text/x-gosu-template"},
-	},
-	Rules{
-		"root": {},
-	},
-))
